requirements.txt
pillow>=9.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
phue>=1.1.0
lifxlan>=1.2.0
yeelight>=0.7.0
matplotlib>=3.5.0
requests>=2.25.0
pyautogui>=0.9.53 